# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  nickname: string;
  score: number;
  levelsCleared: number;
  stars: number;
  isMe: boolean;
}