# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  me: LeaderboardEntry | null;
  around: LeaderboardEntry[];
}