# lib/api-zod/src/generated/types/nicknameRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NicknameRequest {
  /**
     * @minLength 3
     * @maxLength 16
     */
  nickname: string;
}