# lib/api-zod/src/generated/types/registeredPlayer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisteredPlayer {
  id: string;
  secret: string;
  nickname: string;
}