# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreResult {
  accepted: boolean;
  score: number;
  levelsCleared?: number;
  stars?: number;
  rank: number;
  total: number;
}