# lib/api-zod/src/generated/types/scoreSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreSubmission {
  /** @minimum 0 */
  score: number;
  /**
     * @minimum 0
     * @maximum 500
     */
  levelsCleared: number;
  /** @minimum 0 */
  stars: number;
  /** @minimum 1 */
  bestLevel: number;
}